backend/requirements.txt
amqp==5.1.1
anyio==3.6.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
exceptiongroup==1.1.2
Faker==19.2.0
fastapi==0.94.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
kombu==5.2.4
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.38
pydantic==1.10.6
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
vine==5.0.0
wcwidth==0.2.6
